% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_version.R
\name{check_wcvp_version}
\alias{check_wcvp_version}
\title{Check that rWCVP is using the latest version of the World Checklist}
\usage{
check_wcvp_version(silent = FALSE, test = FALSE)
}
\arguments{
\item{silent}{Logical. If FALSE, function prints the version status to
console. If TRUE, function returns either TRUE (if version up to date) or
FALSE (if version not up to date), for use in warnings. Defaults to FALSE.}

\item{test}{Logical. Defaults to FALSE. Only used for testing, should not be changed.}
}
\description{
Check that rWCVP is using the latest version of the World Checklist
}
\examples{
check_wcvp_version()

}
