% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coast}
\alias{coast}
\title{Aggregated WGSRPD3 polygons, for mapping}
\format{
An 'sf_MULTIPOLYGON' object.
}
\source{
Aggregated from: \url{https://github.com/tdwg/wgsrpd/tree/master/level3}
}
\usage{
coast
}
\description{
Spatial data from WGSRPD Level 3, for plotting maps
}
\keyword{datasets}
