% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{wcvp_check_version}
\alias{wcvp_check_version}
\title{Check if the packaged version of WCVP is up to date.}
\usage{
wcvp_check_version(silent = FALSE)
}
\arguments{
\item{silent}{Raise a warning if the version is out of date.}
}
\description{
Check if the packaged version of WCVP is up to date.
}
\examples{
wcvp_check_version()

}
