% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{wcvp_version}
\alias{wcvp_version}
\title{Functions to help check the WCVP data is okay and up to date.
View the version of WCVP available.}
\usage{
wcvp_version(long = TRUE)
}
\arguments{
\item{long}{Whether to return the version date with version number.}
}
\description{
Functions to help check the WCVP data is okay and up to date.
View the version of WCVP available.
}
\examples{
wcvp_version()

}
